Use a platform independent copy task for the built dll
After the dll is built, it is copied to the mod directory. This way the mod
directory contains both the assets and the dll, making it a complete unzipped
mod. The .gitignore file prevents the dll from showing up in the git status,
since the rest of the mod directory is tracked in git.

Previously the csproj was executing the "copy" command to copy the files.
However, Linux does not have a "copy" command. Its equivalent command is "cp".
Now the project uses the platform independent Copy task instead of executing a
shell command.